+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Compatibility:
 * Fix `Coverage.supported?` and raise `TypeError` if argument is not Symbol (#3039, @andrykonchin).
 * Accept options argument to `Regexp.{new,compile}` of String and warn for unknown types (#3039, @rwstauner).
 * Implement `Time#deconstruct_keys` from Ruby 3.2 (#3039, @rwstauner).
+* Handle either positional or keywords arguments by default in `Struct.new` (#3039, @rwstauner).
 
 Performance: